Fix Clang Tidy-Fix Modernisation Errors
Mainly being:

1 - Adding `!= nullptr` into the while loops.
While it's not wrong, the layout should a bit different. Though, reverting changes is easier.

2 - Vector `nullptr` checks.
Probably because of the vec3_t compatibility layer, Clang thinks Vector variables as pointers, which are not.

3 - Repeat of the `override` specifier.
Possibly because of applying Tidy-Fix per project, same files processed by different operation steps and fix are applied by both.
But in any case, `override override` is replaced with `override`